Only pass in immutable optional arguments
As explained in [1], optional arguments in Python only get
instantiated once. This patch fixes weird behavior where
TeXCollections would magically contain objects that had been inserted
into previously-created TeXCollections (often in entirely different
pieces of code). Ultimately this comes from the fact that the default
list that is created ("objs") is mutable, and so adding to it adds to
the globally shared instance.

[1] http://stackoverflow.com/questions/2899643/python-constructor-does-weird-things-with-optional-parameters