Handle the TeXRow ending behavior inside TeXRow.compile()
Previously, we had to handle the TeXRow ending behavior (whether it
ends with \\ or nothing depending on whether it's the last item in a
collection) within addObj of the LaTeX environment. Now that we record
what collection TeXObjects are in and handle the start/end/etc. as
TeXObjects, we can handle this entirely within TeXRow itself.