Let TeXObject accept other non-compile able types other than strings
Currently TeXObject is set up to accept two kinds of objects ---
strings and things that have a .compile() method. This is weird and
unpredictable; this patch makes it explicitly check if an object is an
instance of a subclass of TeXObject and if not, runs str() on it. This
does make it mandatory to subclass TeXObject for new types, but that's
probably a good thing anyway.